Return orbit-interface actions etc. as floats rather than arrays (consistent with other outputs)

+++ galpy/orbit_src/Orbit.py
@@ -1115,9 +1115,9 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return self._orb._aA(self(),use_physical=False)[0]
+            return float(self._orb._aA(self(),use_physical=False)[0])
         else:
-            return self._orb._aA(self,use_physical=False)[0]
+            return float(self._orb._aA(self,use_physical=False)[0])
 
     @physical_conversion('action')
     def jp(self,pot=None,**kwargs):
@@ -1166,9 +1166,9 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return self._orb._aA(self(),use_physical=False)[1]
+            return float(self._orb._aA(self(),use_physical=False)[1])
         else:
-            return self._orb._aA(self,use_physical=False)[1]
+            return float(self._orb._aA(self,use_physical=False)[1])
 
     @physical_conversion('action')
     def jz(self,pot=None,**kwargs):
@@ -1217,9 +1217,9 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return self._orb._aA(self(),use_physical=False)[2]
+            return float(self._orb._aA(self(),use_physical=False)[2])
         else:
-            return self._orb._aA(self,use_physical=False)[2]
+            return float(self._orb._aA(self,use_physical=False)[2])
 
     @physical_conversion('angle')
     def wr(self,pot=None,**kwargs):
@@ -1262,11 +1262,11 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return self._orb._aA.actionsFreqsAngles(self(),
-                                                    use_physical=False)[6][0]
+            return float(self._orb._aA.actionsFreqsAngles(self(),
+                                                    use_physical=False)[6][0])
         else:
-            return self._orb._aA.actionsFreqsAngles(self,
-                                                    use_physical=False)[6][0]
+            return float(self._orb._aA.actionsFreqsAngles(self,
+                                                    use_physical=False)[6][0])
 
     @physical_conversion('angle')
     def wp(self,pot=None,**kwargs):
@@ -1309,11 +1309,11 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return self._orb._aA.actionsFreqsAngles(self(),
-                                                    use_physical=False)[7][0]
+            return float(self._orb._aA.actionsFreqsAngles(self(),
+                                                    use_physical=False)[7][0])
         else:
-            return self._orb._aA.actionsFreqsAngles(self,
-                                                    use_physical=False)[7][0]
+            return float(self._orb._aA.actionsFreqsAngles(self,
+                                                    use_physical=False)[7][0])
 
     @physical_conversion('angle')
     def wz(self,pot=None,**kwargs):
@@ -1356,11 +1356,11 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return self._orb._aA.actionsFreqsAngles(self(),
-                                                    use_physical=False)[8][0]
+            return float(self._orb._aA.actionsFreqsAngles(self(),
+                                                    use_physical=False)[8][0])
         else:
-            return self._orb._aA.actionsFreqsAngles(self,
-                                                    use_physical=False)[8][0]
+            return float(self._orb._aA.actionsFreqsAngles(self,
+                                                    use_physical=False)[8][0])
 
     @physical_conversion('time')
     def Tr(self,pot=None,**kwargs):
@@ -1409,11 +1409,11 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return 2.*nu.pi/self._orb._aA.actionsFreqs(self(),
-                                                       use_physical=False)[3][0]
+            return float(2.*nu.pi/self._orb._aA.actionsFreqs(self(),
+                                                       use_physical=False)[3][0])
         else:
-            return 2.*nu.pi/self._orb._aA.actionsFreqs(self,
-                                                       use_physical=False)[3][0]
+            return float(2.*nu.pi/self._orb._aA.actionsFreqs(self,
+                                                       use_physical=False)[3][0])
 
     @physical_conversion('time')
     def Tp(self,pot=None,**kwargs):
@@ -1462,11 +1462,11 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return 2.*nu.pi/self._orb._aA.actionsFreqs(self(),
-                                                       use_physical=False)[4][0]
+            return float(2.*nu.pi/self._orb._aA.actionsFreqs(self(),
+                                                       use_physical=False)[4][0])
         else:
-            return 2.*nu.pi/self._orb._aA.actionsFreqs(self,
-                                                       use_physical=False)[4][0]
+            return float(2.*nu.pi/self._orb._aA.actionsFreqs(self,
+                                                       use_physical=False)[4][0])
 
     def TrTp(self,pot=None,**kwargs):
         """
@@ -1508,9 +1508,9 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return self._orb._aA.actionsFreqs(self())[4][0]/self._orb._aA.actionsFreqs(self())[3][0]*nu.pi
+            return float(self._orb._aA.actionsFreqs(self())[4][0]/self._orb._aA.actionsFreqs(self())[3][0]*nu.pi)
         else:
-            return self._orb._aA.actionsFreqs(self)[4][0]/self._orb._aA.actionsFreqs(self)[3][0]*nu.pi
+            return float(self._orb._aA.actionsFreqs(self)[4][0]/self._orb._aA.actionsFreqs(self)[3][0]*nu.pi)
  
     @physical_conversion('time')
     def Tz(self,pot=None,**kwargs):
@@ -1559,11 +1559,11 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return 2.*nu.pi/self._orb._aA.actionsFreqs(self(),
-                                                       use_physical=False)[5][0]
+            return float(2.*nu.pi/self._orb._aA.actionsFreqs(self(),
+                                                       use_physical=False)[5][0])
         else:
-            return 2.*nu.pi/self._orb._aA.actionsFreqs(self,
-                                                       use_physical=False)[5][0]
+            return float(2.*nu.pi/self._orb._aA.actionsFreqs(self,
+                                                       use_physical=False)[5][0])
 
     @physical_conversion('frequency')
     def Or(self,pot=None,**kwargs):
@@ -1610,9 +1610,9 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return self._orb._aA.actionsFreqs(self(),use_physical=False)[3][0]
+            return float(self._orb._aA.actionsFreqs(self(),use_physical=False)[3][0])
         else:
-            return self._orb._aA.actionsFreqs(self,use_physical=False)[3][0]
+            return float(self._orb._aA.actionsFreqs(self,use_physical=False)[3][0])
 
     @physical_conversion('frequency')
     def Op(self,pot=None,**kwargs):
@@ -1658,9 +1658,9 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return self._orb._aA.actionsFreqs(self(),use_physical=False)[4][0]
+            return float(self._orb._aA.actionsFreqs(self(),use_physical=False)[4][0])
         else:
-            return self._orb._aA.actionsFreqs(self,use_physical=False)[4][0]
+            return float(self._orb._aA.actionsFreqs(self,use_physical=False)[4][0])
 
     @physical_conversion('frequency')
     def Oz(self,pot=None,**kwargs):
@@ -1706,9 +1706,9 @@ class Orbit(object):
         _check_consistent_units(self,pot)
         self._orb._setupaA(pot=pot,**kwargs)
         if self._orb._aAType.lower() == 'isochroneapprox':
-            return self._orb._aA.actionsFreqs(self(),use_physical=False)[5][0]
+            return float(self._orb._aA.actionsFreqs(self(),use_physical=False)[5][0])
         else:
-            return self._orb._aA.actionsFreqs(self,use_physical=False)[5][0]
+            return float(self._orb._aA.actionsFreqs(self,use_physical=False)[5][0])
 
     def time(self,*args,**kwargs):
         """
